+++ src/lib/hooks/useCandidates.ts
@@ -8,6 +8,7 @@ import {
     createCandidate as createCandidateApi,
     batchCreateCandidates as batchCreateCandidatesApi,
 } from '@/lib/api/positions';
+import { sendAssessmentInvitation } from '@/lib/api/assessments';
 
 interface UseCandidatesReturn {
     candidates: ApplicationDisplayInfo[];
@@ -16,6 +17,8 @@ interface UseCandidatesReturn {
     positionTitle: string | null;
     createCandidate: (candidate: AddApplicationWithCandidateDataDTO) => Promise<void>;
     batchCreateCandidates: (candidates: AddApplicationWithCandidateDataDTO[]) => Promise<void>;
+    isSendingAssessments: boolean;
+    handleSendAssessments: () => Promise<void>;
 }
 
 export default function useCandidates(positionId: string): UseCandidatesReturn {
@@ -23,6 +26,7 @@ export default function useCandidates(positionId: string): UseCandidatesReturn {
     const [loading, setLoading] = useState(true);
     const [error, setError] = useState<string | null>(null);
     const [positionTitle, setPositionTitle] = useState<string | null>(null);
+    const [isSendingAssessments, setIsSendingAssessments] = useState(false);
 
     useEffect(() => {
         if (!positionId) return;
@@ -89,6 +93,38 @@ export default function useCandidates(positionId: string): UseCandidatesReturn {
         }
     };
 
+    const handleSendAssessments = async () => {
+        try {
+            setIsSendingAssessments(true);
+            const result = await sendAssessmentInvitation(positionId);
+
+            if (result.totalSent > 0) {
+                toast.success(
+                    `Successfully sent ${result.totalSent} assessment invitation${result.totalSent !== 1 ? 's' : ''}`
+                );
+            }
+
+            if (result.totalFailed > 0) {
+                toast.error(
+                    `Failed to send ${result.totalFailed} invitation${result.totalFailed !== 1 ? 's' : ''}`
+                );
+            }
+
+            if (result.totalSent === 0 && result.totalFailed === 0) {
+                toast.info('No candidates with pending assessments to send');
+            }
+        } catch (err) {
+            const errorMessage = (err as Error).message;
+            if (errorMessage.includes('does not have an assessment template assigned')) {
+                toast.error('This position does not have an assessment template assigned');
+            } else {
+                toast.error(errorMessage || 'Failed to send assessment invitations');
+            }
+        } finally {
+            setIsSendingAssessments(false);
+        }
+    };
+
     return {
         candidates,
         loading,
@@ -96,5 +132,7 @@ export default function useCandidates(positionId: string): UseCandidatesReturn {
         positionTitle,
         createCandidate,
         batchCreateCandidates,
+        isSendingAssessments,
+        handleSendAssessments,
     };
 }
